+++ xpsservice/models.py
@@ -9,35 +9,8 @@ from ase import Atoms
 from pydantic import BaseModel, Field, validator
 from rdkit import Chem
 
+from .settings import *
 
-
-
-# Define transition_map dictionary
-# Add more entries for other orbitals as needed
-transition_map = {
-    "C1s": {
-        "element": "C",
-        "orbital": "1s",
-        "soap_filepath": os.path.abspath("xpsservice/SOAP_turbo_C1s.txt"),
-        "model_filepath": os.path.abspath("xpsservice/XPS_GPR_C1s_xtb.pkl")
-    },
-    "O1s": {
-        "element": "O",
-        "orbital": "1s",
-        "soap_filepath": os.path.abspath("xpsservice/SOAP_turbo_O1s.txt"),
-        "model_filepath": os.path.abspath("xpsservice/XPS_GPR_O1s_xtb.pkl")
-    },   
-}
-
-
-# Derive allowed elements from transition_map
-def derive_allowed_elements(transition_map: dict) -> Set[str]:
-    allowed_elements = {info["element"] for info in transition_map .values()}
-    return allowed_elements
-
-ALLOWED_ELEMENTS = derive_allowed_elements(transition_map)
-ALLOWED_METHODS = ("GFNFF", "GFN2xTB", "GFN1xTB")
-ALLOWED_FF = ("uff", "mmff94", "mmff94s")
 
 #MM, bug: Not sure to be needed
 class TransitionValidator(BaseModel):
@@ -55,63 +28,6 @@ class OptimizationResult:
     atoms: Atoms
     forces: np.ndarray
     energy: float
-
-
-class IRResult(BaseModel):
-    wavenumbers: List[float] = Field(None, description="List of wavenumbers in cm^-1")
-    intensities: List[float] = Field(
-        None, description="List of IR intensities in (D/Å)^2 amu^-1"
-    )
-    ramanIntensities: List[float] = Field(
-        None,
-        description="List of Raman intensities in (D/Å)^2 amu^-1, computed using Placzek and Bond Polarization (using values from Lippincott/Stuttman) approximation",
-    )
-    zeroPointEnergy: float = Field(None, description="Zero point energy in a.u.")
-    modes: Optional[List[dict]] = Field(
-        None,
-        description="List of dictionaries with the keys `number` - number of the mode (zero indexed), `displacements` - xyz file with the displacement vectors, `intensity` - IR intensity of the mode in D/Å)^2 amu^-1, `ramanIntensity` - Raman intensity of mode, `imaginary` - true if mode is imaginary, `mostDisplaceAtoms` - sorted list of atom indices (zero indiced) according to they displacement (Euclidean norm), `mostContributingAtoms` - most contributing atoms according to a distance criterion.",
-    )
-    mostRelevantModesOfAtoms: Optional[Dict[int, List[int]]] = Field(
-        None,
-        description="Dictionary indexed with atom indices (zero indexed) and mode indices (zero indexed) as values that is most relevant for a given",
-    )
-    mostRelevantModesOfBonds: Optional[List[dict]] = Field(
-        None,
-        description="List of dictionaries with the key `startAtom`, `endAtom` and `mode`",
-    )
-    hasImaginaryFrequency: bool = Field(
-        None, description="True if there is any mode with imaginary frequency"
-    )
-    isLinear: bool = Field(None, description="True if the molecule is linear.")
-    momentsOfInertia: List[float] = Field(
-        None,
-        description="Moments of inertia around principal axes. For a linear molecule one only expects two non-zero components.",
-    )
-    hasLargeImaginaryFrequency: bool = Field(
-        None,
-        description="True if there is a large imaginary frequency, indicating a failed geometry optimization.",
-    )
-
-
-class IRRequest(BaseModel):
-    smiles: Optional[str] = Field(
-        None,
-        description="SMILES string of input molecule. The service will add implicit hydrogens",
-    )
-    molFile: Optional[str] = Field(
-        None,
-        description="String with molfile with expanded hydrogens. The service will not attempt to add implicit hydrogens to ensure that the atom ordering is preserved.",
-    )
-    method: Optional[str] = Field(
-        "GFN2xTB",
-        description="String with method that is used for geometry optimization and calculation of the vibrational frequencies. Allowed values are `GFNFF`, `GFN2xTB`, and `GFN1xTB`. `GFNFF` is the computationally most inexpensive method, but can be less accurate than the xTB methods",
-    )
-
-    @validator("method")
-    def method_match(cls, v):
-        if not v in ALLOWED_METHODS:
-            raise ValueError(f"method must be in {ALLOWED_METHODS}")
-        return v
 
 
 class ConformerRequest(BaseModel):
@@ -191,6 +107,10 @@ class XPSRequest(BaseModel):
         "GFN2xTB",
         description="String with method that is used for geometry optimization and calculation of the XPS binding energies. Allowed values are `GFNFF`, `GFN2xTB`, and `GFN1xTB`. `GFNFF` is the computationally most inexpensive method, but can be less accurate than the xTB methods",
     )
+    fmax: Optional[float] = Field(
+        0.01,
+        description="Maximum force admissible during the geometry optimization process using either of the selected method. Typically ranging from 0.1 to 0.0001 "
+    )
 
     @validator("smiles")
     def validate_smiles(cls, v):
@@ -222,6 +142,13 @@ class XPSRequest(BaseModel):
     @validator("method")
     def validate_method(cls, v):
         if v not in ALLOWED_METHODS:
+            raise ValueError(f"Method must be in {ALLOWED_METHODS}")
+        return v
+    
+    # Validator for the method field to ensure the method is within the allowed list of methods
+    @validator("fmax")
+    def validate_method(cls, v):
+        if v not in ALLOWED_FMAX:
             raise ValueError(f"Method must be in {ALLOWED_METHODS}")
         return v
